music module content merged into core.Sound class
